+++ packages/terra-date-time-picker/tests/jest/DateTimePicker.test.jsx
@@ -50,6 +50,12 @@ it('should render a date time picker with requestFocus', () => {
   expect(datePicker).toMatchSnapshot();
 });
 
+it('should render a date time picker with onBlur', () => {
+  const onBlur = () => {};
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" onBlur={onBlur} utcOffset={0} />);
+  expect(datePicker).toMatchSnapshot();
+});
+
 it('should render a date time picker with onChange', () => {
   const onChange = () => {};
   const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" onChange={onChange} utcOffset={0} />);
@@ -64,6 +70,13 @@ it('should render a date time picker with onChangeRaw', () => {
 
 it('should render a date time picker with a default date and time', () => {
   const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" value="2017-04-01T10:30" utcOffset={0} />);
+  expect(datePicker).toMatchSnapshot();
+});
+
+it('should render with a default timeVariant if variant is not specified', () => {
+  const defaultTimeVariant = '24-hour';
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" value="2017-04-01T10:30" utcOffset={0} />);
+  expect(datePicker.prop('timeVariant')).toBe(defaultTimeVariant);
   expect(datePicker).toMatchSnapshot();
 });