+++ packages/terra-date-time-picker/tests/wdio/date-time-picker-spec.js
@@ -52,8 +52,8 @@ describe('DateTimePicker', () => {
   describe('Time Clarification Dialog', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-dst');
-      browser.click('input[name="terra-time-hour-input"]');
       browser.click('input[name="terra-time-minute-input"]');
+      browser.keys('Tab');
       browser.waitForVisible('[class*="time-clarification"]');
     });
 
@@ -75,8 +75,8 @@ describe('DateTimePicker', () => {
   describe('Time Clarification Dialog Dismissal', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-dst');
-      browser.click('input[name="terra-time-hour-input"]');
       browser.click('input[name="terra-time-minute-input"]');
+      browser.keys('Tab');
       browser.waitForVisible('[class*="time-clarification"]');
     });
 
@@ -93,8 +93,8 @@ describe('DateTimePicker', () => {
     if (includesTZ) {
       before(() => {
         browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-dst');
-        browser.click('input[name="terra-time-hour-input"]');
         browser.click('input[name="terra-time-minute-input"]');
+        browser.keys('Tab');
         browser.waitForVisible('[class*="time-clarification"]');
       });
 
@@ -112,8 +112,8 @@ describe('DateTimePicker', () => {
     if (includesTZ) {
       before(() => {
         browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-dst');
-        browser.click('input[name="terra-time-hour-input"]');
         browser.click('input[name="terra-time-minute-input"]');
+        browser.keys('Tab');
         browser.waitForVisible('[class*="time-clarification"]');
       });
 
@@ -130,8 +130,8 @@ describe('DateTimePicker', () => {
   describe('Time Clarification Dialog Disabled', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-dst-disabled');
-      browser.click('input[name="terra-time-hour-input"]');
       browser.click('input[name="terra-time-minute-input"]');
+      browser.keys('Tab');
       browser.waitForVisible('[class*="time-clarification"]');
       browser.click('[class*="button-daylight"]');
       browser.click('#date-time-picker-toggler');
@@ -139,6 +139,74 @@ describe('DateTimePicker', () => {
 
     const ignoredDisabledAlly = Object.assign({ 'color-contrast': { enabled: false } }, ignoredA11y);
     Terra.should.beAccessible({ rules: ignoredDisabledAlly });
+    Terra.should.matchScreenshot({ viewports });
+  });
+
+  describe('OnBlur', () => {
+    before(() => {
+      browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-focus-blur');
+    });
+
+    it('puts focus on the input', () => {
+      browser.click('input[name="terra-date-input"]');
+      // Removes the blinking cursor to prevent screenshot mismatches.
+      browser.execute('document.querySelector(\'input[name="terra-date-input"]\').style.caretColor = "transparent";');
+      expect(browser.getText('#blur-count')).to.equal('0');
+      expect(browser.getText('#focus-count')).to.equal('1');
+      expect(browser.getText('#iso')).to.equal('');
+      expect(browser.getText('#input-value')).to.equal('');
+      expect(browser.getText('#complete-date')).to.equal('No');
+      expect(browser.getText('#valid-date')).to.equal('Yes');
+    });
+
+    it('enter a complete date to move focus to the hour input and onBlur is not triggered', () => {
+      browser.keys('05/01/2019');
+      expect(browser.getText('#blur-count')).to.equal('0');
+      expect(browser.getText('#focus-count')).to.equal('1');
+      expect(browser.getText('#iso')).to.equal('');
+      expect(browser.getText('#input-value')).to.equal('');
+      expect(browser.getText('#complete-date')).to.equal('No');
+      expect(browser.getText('#valid-date')).to.equal('Yes');
+    });
+
+    it('enter the hour to move focus to the minute input and onBlur is not triggered', () => {
+      browser.setValue('input[name="terra-time-hour-input"]', '10');
+      expect(browser.getText('#blur-count')).to.equal('0');
+      expect(browser.getText('#focus-count')).to.equal('1');
+      expect(browser.getText('#iso')).to.equal('');
+      expect(browser.getText('#input-value')).to.equal('');
+      expect(browser.getText('#complete-date')).to.equal('No');
+      expect(browser.getText('#valid-date')).to.equal('Yes');
+    });
+
+    it('enter the minute and onBlur is not triggered', () => {
+      browser.setValue('input[name="terra-time-minute-input"]', '10');
+      expect(browser.getText('#blur-count')).to.equal('0');
+      expect(browser.getText('#focus-count')).to.equal('1');
+      expect(browser.getText('#iso')).to.equal('');
+      expect(browser.getText('#input-value')).to.equal('');
+      expect(browser.getText('#complete-date')).to.equal('No');
+      expect(browser.getText('#valid-date')).to.equal('Yes');
+    });
+
+    it('tabs out of the component and onBlur is triggered', () => {
+      browser.keys('Tab');
+      expect(browser.getText('#blur-count')).to.equal('1');
+      expect(browser.getText('#iso')).to.equal('2019-05-01T10:10:00-05:00');
+      expect(browser.getText('#input-value')).to.equal('05/01/2019 10:10');
+      expect(browser.getText('#complete-date')).to.equal('Yes');
+      expect(browser.getText('#valid-date')).to.equal('Yes');
+    });
+  });
+
+  describe('Valid date entry moves focus to hour input', () => {
+    before(() => {
+      browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-focus-blur');
+      // Removes the blinking cursor to prevent screenshot mismatches.
+      browser.execute('document.querySelector(\'input[name="terra-time-minute-input"]\').style.caretColor = "transparent";');
+      browser.setValue('input[name="terra-date-input"]', '04/15/2019');
+    });
+
     Terra.should.matchScreenshot({ viewports });
   });
 
@@ -231,22 +299,6 @@ describe('DateTimePicker', () => {
     Terra.should.matchScreenshot({ viewports });
   });
 
-  describe('Launches and Dismisses modal manager after date selection', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-inside-modal');
-      browser.click('[class*="disclose"]');
-      browser.click('[class*="custom-input"] > [class*="button"]');
-    });
-
-    Terra.should.matchScreenshot('0', { viewports });
-
-    it('Dismissed the modal manager after a click', () => {
-      browser.click('[aria-label="day-27"]');
-    });
-
-    Terra.should.matchScreenshot('1', { viewports });
-  });
-
   describe('OnSelect', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-on-select');
@@ -268,61 +320,13 @@ describe('DateTimePicker', () => {
     Terra.should.matchScreenshot('2', { viewports });
   });
 
-  describe('Displays the date picker inside a modal manager and exits when Enter is pressed', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-inside-modal');
-      browser.click('[class*="disclose"]');
-      browser.click('[class*="custom-input"] > [class*="button"]');
-    });
-
-    Terra.should.matchScreenshot('0', { viewports, selector: '[class*="abstract-modal"]' });
-
-    it('Dismissed the modal manager after a click', () => {
-      browser.keys(['Enter']);
-    });
-
-    Terra.should.matchScreenshot('1', { viewports });
-  });
-
-  describe('Displays the date picker inside a modal manager and exits when Escape is pressed', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-inside-modal');
-      browser.click('[class*="disclose"]');
-      browser.click('[class*="custom-input"] > [class*="button"]');
-    });
-
-    Terra.should.matchScreenshot('0', { viewports, selector: '[class*="abstract-modal"]' });
-
-    it('Dismissed the modal manager after a click', () => {
-      browser.keys(['Escape']);
-    });
-
-    Terra.should.matchScreenshot('1', { viewports });
-  });
-
-  describe('Displays the date picker inside a modal manager and exits when Tab is pressed', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-inside-modal');
-      browser.click('[class*="disclose"]');
-      browser.click('[class*="custom-input"] > [class*="button"]');
-    });
-
-    Terra.should.matchScreenshot('0', { viewports, selector: '[class*="abstract-modal"]' });
-
-    it('Dismissed the modal manager after a click', () => {
-      browser.keys(['Tab']);
-    });
-
-    Terra.should.matchScreenshot('1', { viewports });
-  });
-
   describe('Clears the default date and time on the calendar button click when default date is excluded', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-default-date-excluded');
       browser.click('[class*="button"]');
-    });
 
-    Terra.should.matchScreenshot({ viewports });
+      expect(browser.getText('#date-time-value')).to.equal('');
+    });
   });
 
   describe('Clears the default date and time on date input focus when default date is excluded', () => {
@@ -384,6 +388,14 @@ describe('DateTimePicker', () => {
       // Removes the blinking cursor to prevent screenshot mismatches.
       browser.execute('document.querySelector(\'input[name="terra-time-minute-input"]\').style.caretColor = "transparent";');
       browser.click('input[name="terra-time-minute-input"]');
+    });
+
+    Terra.should.matchScreenshot({ viewports });
+  });
+
+  describe('Time Input displays Twelve Hour format when timeVariant is specified', () => {
+    before(() => {
+      browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-time-input-attributes');
     });
 
     Terra.should.matchScreenshot({ viewports });